Working machine learning

+++ ChartAutoRating/DataSample.cs
@@ -2,9 +2,9 @@
     public readonly struct DataSample {
         public int DifficultyRating { get; }
             
-        public float[] Metrics { get; }
+        public double[] Metrics { get; }
 
-        public DataSample(int difficultyRating, float[] metrics) {
+        public DataSample(int difficultyRating, double[] metrics) {
             DifficultyRating = difficultyRating;
             Metrics = metrics;
         }
